Features: - Engine: Particle Effects #99
Fixes:
- Embedded assets now can be picked in AssetSelect
- Compilation warnings fixed
- Singleton template removed from Next
- Other minor fixes

+++ engine/src/timer.cpp
@@ -1,12 +1,12 @@
 #include "timer.h"
 
 static TimePoint m_sLastTime;
-static double m_sFixedDelta = 0.0;
-static double m_sTime       = 0.0;
-static double m_sDeltaTime  = 0.0;
-static double m_sTimeScale  = 1.0;
+static float m_sFixedDelta  = 0.0;
+static float m_sTime        = 0.0;
+static float m_sDeltaTime   = 0.0;
+static float m_sTimeScale   = 1.0;
 
-void Timer::init(double fixed) {
+void Timer::init(float fixed) {
     m_sLastTime     = std::chrono::high_resolution_clock::now();
     m_sFixedDelta   = fixed;
 }
@@ -14,27 +14,27 @@ void Timer::init(double fixed) {
 void Timer::update() {
     TimePoint current   = std::chrono::high_resolution_clock::now();
 
-    m_sDeltaTime    = (std::chrono::duration_cast<std::chrono::duration<double> >(current - m_sLastTime)).count();
+    m_sDeltaTime    = (std::chrono::duration_cast<std::chrono::duration<float> >(current - m_sLastTime)).count();
     m_sTime        += m_sDeltaTime;
     m_sLastTime     = current;
 }
 
-double Timer::time() {
+float Timer::time() {
     return m_sTime;
 }
 
-double Timer::deltaTime() {
+float Timer::deltaTime() {
     return m_sDeltaTime;
 }
 
-double Timer::fixedDelta() {
+float Timer::fixedDelta() {
     return m_sFixedDelta;
 }
 
-double Timer::scale() {
+float Timer::scale() {
     return m_sTimeScale;
 }
 
-void Timer::setScale(double scale) {
+void Timer::setScale(float scale) {
     m_sTimeScale    = scale;
 }
